security(repo): separate write authority from tool acknowledgement
